fetch BE api for realEstate, bankSaving
