+++ src/services/external-service/index.ts
@@ -1,2 +1,2 @@
-import { coinGeckoService, finhubService } from 'services';
-export {coinGeckoService,finhubService}
+export * from './coingecko-service';
+export * from './finhub-service';
