+++ src/services/firebase-service/index.ts
@@ -1,4 +1,2 @@
-import { facebookAuth } from './facebookAuth';
-import { googleAuth } from './googleAuth';
-
-export { facebookAuth, googleAuth };
+export * from './facebookAuth';
+export * from './googleAuth';
