+++ src/services/index.ts
@@ -1,20 +1,7 @@
-import { facebookAuth, googleAuth } from 'services/firebase-service';
-import { cryptoService } from './crypto-service';
-import { storageService } from './storage-service';
-import { alertService } from './alert-service';
-import { userService } from './user-service';
-import { httpService } from './http-service';
-import { coinGeckoService } from './external-service/coingecko-service';
-import { finhubService } from './external-service/finhub-service';
-
-export {
-  coinGeckoService,
-  finhubService,
-  httpService,
-  userService,
-  storageService,
-  alertService,
-  cryptoService,
-  facebookAuth,
-  googleAuth,
-};
+export * from './external-service';
+export * from './firebase-service';
+export * from './alert-service';
+export * from './crypto-service';
+export * from './http-service';
+export * from './storage-service';
+export * from './user-service'
