+++ src/store/assets-detail.store/index.ts
@@ -1,4 +1,2 @@
-import { cryptoVolatilityDetailStore } from './crypto-volatility-detail.store';
-import { stockVolatilityDetailStore } from './stock-volatility-detail.store';
-
-export { cryptoVolatilityDetailStore, stockVolatilityDetailStore };
+export * from './crypto-volatility-detail.store';
+export * from './stock-volatility-detail.store';
