+++ src/store/assets-detail.store/stock-volatility-detail.store.ts
@@ -1,6 +1,6 @@
 import { action, computed, makeAutoObservable, observable } from 'mobx';
 import { finhubService } from 'services';
-import { portfolioData } from 'store/portfolio-detail.store';
+import { portfolioData } from 'store/portfolio/portfolio-data';
 
 class StockVolatilityDetailStore {
   isOpenAddNewTransactionModal: boolean = false;
