+++ src/store/index.ts
@@ -1,15 +1,4 @@
-import { userStore } from './user.store';
-import { authStore } from './auth.store';
-import { portfolioDetailStore } from './portfolio-detail.store';
-import {
-  stockVolatilityDetailStore,
-  cryptoVolatilityDetailStore,
-} from './assets-detail.store';
-
-export {
-  authStore,
-  userStore,
-  portfolioDetailStore,
-  cryptoVolatilityDetailStore,
-  stockVolatilityDetailStore,
-};
+export * from './user.store';
+export * from './assets-detail.store';
+export * from './auth.store';
+export * from './portfolio';
