+++ src/store/portfolio/index.ts
@@ -1,0 +1,1 @@
+export * from './portfolio-detail.store';
