+++ src/types/portfolio-detail.model.ts
@@ -15,3 +15,29 @@ export type PortfolioAllocation = {
   symbol: string;
   portfolioAllocation: string;
 };
+
+export type RealEstateItem = 
+{
+  id:number,
+  name:string,
+  inputDay:string,
+  inputMoneyAmount:number,
+  inputCurrency:string,
+  lastChanged:string,
+  description:string,
+  buyPrice:number,
+  currentPrice:number,
+}
+
+export type BankSavingItem = {
+  id:number,
+  name:string,
+  inputDay:string,
+  inputMoneyAmount:number,
+  inputCurrency:string,
+  lastChanged:string,
+  description:string,
+  interestRate:number,
+  termRange:number,
+  isGoingToReinState:string
+}
